Three_Plot_Types_and_AML_Data
Updated the app to work with a sample AML dataset. The app has been updated to display any of the following simultaneously: UMAP plot, feature plot, or violin plot. The user is able to choose which combination of plots they would like with checkboxes. The plots are displayed or hidden using conditionalPanel() UI elements. Input is currently inflexible as group.by and split.by choices currently apply to all graphs: individualized settings will be specified in future versions.