Added Dot Plot
Added ability to create a dot plot. Since the dot plot works best with many features, I created a separate set of feature checkboxes for this plot. All features in the sample set are selected initially. Added group.by and split.by arguments for the dot plots, but split.by doesn't appear to work as expected (no division of the groups by the split.by variable).