Functional Text Entry
Successfully implemented the text entry feature for the feature plot. The text box input is isolated to avoid the plot rendering each time the user types a character in; instead, the app waits until an "update" button is clicked to display the feature. The plot still responds to all changes changes in the split.by argument, or the "make feature plot" checkbox.

A conditional UI is rendered in the server function, based on whether a text value is entered, and whether the feature is valid (for now, valid features are the features accessed by calling rownames(uhg)). The app will either display a plot if the feature is valid, or it will notify the user if an invalid option is chosen, or if no features have been entered.