Text Entry for all Plots; Option to use Separate Text Entry for Dot Plots
The text feature entry has been updated to apply to all plots (added functionality for the dot plot). To allow the user to add more options to the dot plot without cluttering the other plot types, a checkbox has been added under dot plot-specific options that gives the user the choice to use separate features for the dot plot. If the user selects this option, a separate text box will appear, and the dot plot will render only according to the features in this text box. Updateselectizeinput() functions inside observeEvent() functions are used to update the dot plot text entry in the background to make it equal to the "global" features entered; this occurrs while the checkbox to use different features for the dot plot is unchecked, ensuring that the features in the new text entry are available for modification by the user as soon as the box is checked.