User-Controlled Width and Height - Functional for UMAP
Added feature for user to manually specify plot heights if desired (indicated by selecting a checkbox).

Width and height can be adjusted by using either a slider input or a text input. The textbox uses a special input type from the shinyWidgets package that delays reactivity until the user presses the enter key with the textbox selected. The slider and the textbox sync with each other each time the user modifies either input, allowing the user to choose which input to use to adjust dimensions.