User Plot Size Control for all Plot Types
The textbox-slider combos have been added to all plot types, allowing the user to control the size of each plot. A scrollbar was also added to the sidebar to improve accessibility, and the navbar changed to a fixed position (as opposed to the default static layout which disappears when the user scrolls down from the top).