ADT features and Download Buttons
ADT features and metadata columns are now searchable in the feature selectize input boxes. Download buttons have been added for each plot type and allow for each plot to be downloaded individually. Aesthetic improvements have also been made, including limiting the sidebar and main panel size to the size of the window, and adding scrollbars to allow user to access the content without having to scroll back up to access the sidebar.