Toggle Legend
Added the ability to toggle display of the legend on all plot types with a checkbox. The implementation works on all single-panel plots and on multi-panel UMAP plots, but abnormal behavior was observed for feature plots with multiple panels: only the legend on the last (bottom right) feature plot was removed. For violin plots with multiple features, the legend is successfully removed, but when it is present it will appear centered to the last violin plot, which is suboptimal for publication-quality figures.