Update: Set Original Axis Limits on UMAP
The two eventReactive() expressions for computing the plot were condensed back into one to match the structure for the feature plot expressions. The feature plot expressions were condensed into one expression in a previous commit to avoid an error with the feature plots where the plot would not update with the subset due to an error that occurred in the second eventReactive() function.

This commit still contains the bug mentioned in the last commit.