Add Interface for Customizing Correlation Table Plots
A collapsible panel was added to the correlations tab sidebar, which displays when the scatterplot is rendered and contains inputs to change the split.by and group.by variables, and a download button for the scatterplot.

The JavaScript for the collapsible panels was also modified to get the opening/closing behavior to apply to all instances of the panel, regardless of when they were created. Previously, only the panels present at startup responded to user clicks.