Improved Error Handling
Updated the code for all three tabs to use the functions in "./R/error_handling.R". The functions check the error message against a list of pre-defined error messages, and display notifications based on the error returned. The notifications are the same as before, but the code in the tryCatch() is much more streamlined, and the app no longer crashes when a subset with no cells is selected.