First Server Function
To facilitate generalization of the app to arbitrary Seurat objects, I turned the subset stats computation into a server function.

The functions work but are still specific to the metadata columns in the current object. I have decided to focus on the use of modules, which should make the app easier to adapt to different metadata columns. The subset stats outputs in the corr and DGE tabs also collide with one another, and modules will prevent this from being an issue by putting the outputs in different namespaces.