make_subset function
Created a function to compute a subset using the currently defined input criteria. The function works for the correlations tab, but it does not yet work with the DGE tab due to the complexity of the DGE subset code. The subset code also doesn't work in the plots tab since the input variables for the subset criteria have a "plots_" prefix. This operation seems to be a good candidate for a module, so unique inputs can be used on different tabs without namespace clashes.