UI for Assay Options
Created UI that displays the current assays in the Seurat object in a multi-select input. For each selected assay, a card appears prompting the user to enter the assay key and a label for the assay. A checkbox to display the assay on plots is presented: if this is checked, a suffix with the assay label will be generated and displayed on plots and in the feature selection search box.