Update Config File Format
For the metadata config list, the `groups` element is defined even when the user does not specify the use of metadata groups. This is undesirable for processing of the file in the main app. `groups` will now be equal to NULL when the user does not specify the use of groups.