DGE Test Selection Module Complete
The `dge_test_selection` module for displaying and processing inputs regarding the test to perform (mode, group by category, and marker classes or DE groups) is complete. The module avoids namespace collisions while simplifying the code for processing inputs. The output of the module is temporarily printed to screen in the app to see the values being returned.