DGE Tab: Updates to Subset Selection Module
Updated the subset selection module to respond to choices made in the test selections module. When a group by metadata category is chosen from the test selections module, the category is passed to the subset selections module, and the menu corresponding to that category is hidden. The server value for that menu is also excluded from the module output to avoid bugs when subsetting.