Consistent multi-dataset format between modules
While modifying the correlations tab to support multiple datasets, I discovered that the most effective use of modules is to create separate instances of the entire correlations tab module for each dataset, as opposed to creating separate instances of the modules within each tab for each dataset (method used in other tabs). Issues with responsiveness of the subset selections module to user selections were observed in the plots and DGE tabs (separate from the flickering issue); to solve these issues, I updated both tabs to use the format in the correlations tab.