Store seurat objects in session$userData
The Seurat object and the subsets are now stored in session$userData instead of within module environments. Moderate improvements in memory consumption were observed when the plots tab was the only tab used, but memory consumption still increased after changing objects when subsets in the dge and correlations tabs were created.