Limit Unnecessary Module Creation
Added a reactiveValues list that stores the keys of datasets for which modules have previously been created. The observers that create server module instances for each tab have been updated with conditional statements, so that they will only create a new instance of a module when it represents a dataset that has not yet been loaded in the current session.

This has improved memory usage, but the DGE tab now does not respond when switching from the longitudinal dataset from the pheresis dataset. The DGE tab works without issues for the pheresis dataset.