Restore structure of DGE and correlations subsets
Storing subsets in session$userData resulted in the reactive expressions for the subsets in the correlations tab and the DGE tab not forming the correct reactive dependencies when switching datasets, which resulted in the app not responding properly to the submit button in the DGE tab. The original means of storing the subset has been restored to correct this error.