Manual Subset String Entry
Added ability to manually define a subset by entering subset code as a string. On all three tabs, the user may now check a box to use string subsetting; when checked, a block text entry will appear, and the user may enter R code to manually define a subset. The subset should be a conditional statement that will evaluate to TRUE or FALSE, and will be evaluated with any criteria entered in the menus above (cells where the intersect of criteria from the menus and criteria from the string evaluate to TRUE will be selected by the subset).