Updated Auto-Dictionary R Markdown
Added a tab to the auto-generated object dictionary to display statistics for features in the object. I integrated Shiny with R Markdown to create a widget in the new tab with a search bar for features, and a renderUI() statement that displays statistics for the feature entered in the search bar.