Functional Feature Search in App
The object dictionary and feature statistics are now accessible in a dropdown menu after pressing a button with a search icon in the advanced subsetting panel.
Clicking the search button opens a drop-up menu where features may be entered to view summary statistics. A link to the object dictionary and string subsetting help is also presented in the dropdown.