Custom Title Entry for DimPlots
For DimPlots, the user may now choose to use a default title (uses the label from the config file), no title, or a custom title. If a custom title is desired, a text box appears for custom entry. The text box automatically updates when the group by category is changed, and has a reset button which will restore the default value when clicked.