Rename all "UMAP" references to "DimPlot"
Given that multiple reductions are now visible, it is no longer accurate to refer to UMAP plots as such. "UMAP" has been renamed to "DimPlot" in all instances it appears in the code, UI, and comments in the plots tab. UMAP references in the dge tab are left the same for now as only the UMAP projection is currently used.