Set order of metadata in config app
Allow user to choose the order of metadata groups in the config app by using a new drag-and-drop menu. The config file will reflect the changes made, though the UI for the options "cards" does not currently change in response to the selections, which may be confusing to the user.