Bug Fix: Loading Metadata Info from Config File
When the drag-and-drop input to sort metadata in the config app was added, a bug related to the loading of the config file was introduced. When the config file is loaded, the UI in the metadata tab is re-created based on the metadata selected and the order in which it was selected; normally the options entered for each metadata category are preserved when the UI is re-created, but when loading the config file this does not happen, and the options appear blank even when they are defined in the loaded config file.

This was fixed using the JavaScript function `insertAfter`, which moves the element on the page instead of deleting and re-creating it, as is the case with Shiny's `renderUI()`. I created a JavaScript function using `insertAfter`, and integrated it with shiny using `shinyjs::extendShinyjs()`.