Fix Bug with Threshold Picker Interface
A bug was discovered with the threshold picker interface that crashed the main app. The bug was corrected by adding req() statements to the observers involved with defining x-axis limits, to keep conditionals from running before a ridge plot is produced (when a feature is specified in the threshold picker server).