Remove deleted module from source statement (#107)
The module "xlim_modal.R" was added to the `source()` statement in the config app since the threshold_picker module formerly took a dependency on it. xlim_modal was removed from the threshold picker module in #106, but it was not also removed from the `source()` statement; this caused the config app to crash. The issue is corrected here.