Creation of data directory
Created data/ directory with a sample YAML config file. Future iterations of the app should have objects in the data/ directory in order to be selectable in the config app, since the file browser launched from fileinput won't (and probably shouldn't) allow the user to browse all the files on the server.