Adapt config app to respond to reactive object
As a result of user choice of the object through the modal structure, the object is now a reactive variable. The config app must be updated to accommodate this change. Each of the tabs in the app has been converted to a module: this will make it easier to avoid namespace collisions when another object with similar assay/metadata column names is loaded.