Create .gitignore
I originally ignored .gitignore when I created the repository, which caused major issues when I switched back to this branch from the current one. There was no .gitignore on this branch, which caused all of the Seurat objects and other files I added since to be considered changes. When switching to the main branch and opting to "leave changes on this branch", everything that should have been ignored was erased when switching back to main. The files re-appeared when returning to this branch, but they must be ignored to prevent them being erased again when switching back.