Fix bug when loading dataset with no preview type
When loading a dataset without a preview image or plot, the app crashes. The cause was an error in defining the preview section of the config file when no preview content is selected: it was being set to NULL instead of a list with `type = "none"` as an argument. The main app was testing for the value of config$preview$type, which was undefined. This has been corrected, and checks for the existence of config$preview$type were added to the main app.