Removed ability to add image
Datasets with an image preview are not loading properly in the main app due to an error with the path saved based on the fileInput. Due to the challenges associated with defining an image for the preview (the preview will break if the image is moved), and the sufficiency of using a plot, the image preview option has been removed for now.