Fix #152
When adding an ADT threshold, the selectize input menus are updated with the ADTs that are not currently in the threshold table. This should also happen when ADT features are deleted from the table, but the code to achieve this was not present in the deletion section. The code has been added here, and features now re-populate as expected when they are deleted from the table.