Change behavior of launch.browser in app function
Update run_scExploreR.R to use the default option for launch.browser if this argument is not defined. The default behavior is to launch an R Studio browser if the function is ran from R Studio: this will not happen if launch.browser is either TRUE or FALSE (when TRUE, the app will launch in the system's default browser).