Resolve issue with loading config file
The issue in #192 occurs because the code to render the metadata and reductions selection menus only responds to changes in module_data reactive variables (module_data$metadata_selected, module_data$metadata_not_selected, and so on), and these variables do not change the second time a config file is loaded. To resolve this, reactive triggers were added to programmatically trigger updates for the menus each time the config file is loaded.