Fix issue in #197 with error not displaying
No error message was returned because of recent changes to the icon_notification_ui() function. Additional named arguments were added, but they were placed before the ellipsis instead of after. This means that the notification elements passed as variable length arguments passed to the function were being incorrectly interpreted as values for the new arguments, causing notification text to not appear