Correct 404 error with auto dictionary
The 404 error was caused by specifying an improper path when creating the Rmd for the object dictionary. The object dictionary was being created in a folder called "resources" in whatever directory the user was launching the app from, and this folder was not on the resource path. The link to the auto dictionary used "resources/", but for links this was not interpereted as "./resources"; instead it pointed to "inst/www" *in the package code* (this was set up via addResourcePath and works for loading CSS, JS, and the static html pages). Files that change based on the object can't be placed in the package directory.

The problem was solved by creating a new resource path, "session_files", pointing to a temp directory in the local file system of the server (created via tempdir()). The auto dictionary was rendered and saved to tempdir()/auto_dictionary_{session_id}.html, and the hyperlink to the file was set as "session_files/auto_dictionary_{session_id}.html".