Plots tab performance improvement
While testing the performance of the plots tab using the reactlog, it was discovered that reactive expressions were computing for plots even when the plots weren't currently being requested by the user (i.e. the user requested just a feature plot, but reactives for the ridge plot and dot plot were running). This added considerable processing time to the app when an object with 200k cells was used. This was fixed by adding req() statements to reactive expressions in the plot module which respond to the switch for each plot, ensuring that reactive expressions only compute when the user enables a plot via the switches.

If the user only has one or a few plots open, the time to render each plot is now decreased. Plots now re-renders if the user hides a plot then shows it again, but this does not take long.