Adapt code in dge_tab
Modified the code in the dge_tab to accept Seurat or SingleCellExperiment objects. For creating new metadata, the following procedure was agreed upon: the metadata table is first pulled from the object via an S3 method, then the table is modified separately, before being saved to the object via another S3 method.