Loading code for SingleCellExperiment objects
Add the ability to load HDF5-enabled SingleCellExperiment objects into the config app. To load an object, the user must set is_HDF5SummarizedExperiment to TRUE, and set object_path to point to the directory of the HDF5-enabled object (the same directory created when saving the object via saveHDF5SummarizedExperiment).