Fix error with loading of config file
The error with loading the config file was traced back to to the reactive code that updates the preview dimplot in the preview_dimplot module. When a file is loaded where a preview dimplot has a split_by value of "none", this value is passed to ncol_settings and then to n_unique, which results in an error. To keep "none" from being passed to ncol_settings, the conditional statements that trigger input updates based on the information in a config file were updated.