Adapt make_subset to SingleCellExperiment objects
The subset() in make_subset did not work properly for SingleCellExperiment objects (#225). To correct this, a method was added to call subset() with the proper parameters based on the object type (uses "select" instead of "subset" for SingleCellExperiment objects). Additional object-specific code in make_subset was adapted, and the subsetting function now works with SingleCellExperiment objects.