Correct error with re-ordering of plot groups in SingleCellExperiment objects
The error described in #226 was caused by the order of objects in the eventExpr parameter of eventReactive statements, defining what the eventReactive statements should re-compute in response to. When object() is first in the eventExpr statement, the statements will return an error, but no error is returned if object() is placed in another position. This is likely due to SingleCellExperiment objects having a method for c(), which results in R attempting to coerce the other objects in the eventExpr statement to SingleCellExperiment objects when the object is listed first. #226 also occurred in the eventReactive statements defining the UI for the DGE and correlation tabs, so the  position of object() is corrected in these statements as well.