Fix issue with arrows appearing on buttons
After adding a set of radio buttons, it was observed that a triangle would appear to the left of the selected button. This was found to be due to the CSS for the collapsible panel: a rule for the button in active headers was not specific enough, and was being applied to all buttons that are in the "active" state. The CSS is specified here to apply only to buttons with the "collapsible" class.