Editing interface for categorical filters
When the editing button is pressed on a categorical filter, an interface similar to the addition interface appears. The user may change the values of the metadata variable being entered, but not the metadata variable itself. The information is properly saved to the filters list when the confirm button is pressed.