Update ranges in the threshold/range picker
Update code of threshold/range picker module to change it's state to reflect a previously set range when it is loaded into the widget. In the subset filter interface, this will occur when the user presses the edit button next to a filter that is a numeric range. Existing code to update based on a threshold was adapted to behave appropriately based on whether the value entered is a threshold or a range.