Threshold picker: separate plot update code
Previously, the code in the threshold picker widget responding to either clicking the plot or loading a previously defined value would execute the plotting code within separate reactive expressions responding to the click event, or the loading of data. The result was that the code for drawing lines on the plot in response to each event was duplicated between expressions, and the execution of the plotting code was entirely dependent on the containing reactive expressions. The plotting code has been moved to a separate obsserve({}) statement, so it can be written only once and triggered whenever the state variable storing the chosen threshold/range is updated.