Subset filter selections: clear variables when finished editing
ReactiveValues objects corresponding to editing are now cleared when pressing the confirm or cancel buttons while editing. This ensures that the reactive statements in the threshold picker always run when the edit button is pressed next to a feature, and that the plot being edited always reflects the feature being chosen for editing.