Separate req() statements in threshold_picker
An error was observed in the threshold_picker widget in a case where module_data$threshold_x was undefined. A req statement was placed to stop the observe() statement from computing in this case, but the req statement was based on a vector of two variables that could be truthy when one is defined but not the other. To require both variables be defined, the statement has been split into two separate req() statements, either of which will stop the observer if the corresponding variable is undefined.