Update module-subset_selections.R
Fixed a bug that occurs when pressing the edit button for an exiting categorical filter after opening the interface to add a categorical filter. When the edit button is pressed, the interface updates to reflect the new variable, but the selection menu remains set to the variable that was selected when adding the feature. When the confirm button is pressed, the filter being edited will reflect the variable that was being added, while the values included will reflect the variable that was being edited, resulting in an impossible subset.

This is corrected by using the variable name from the editing_data reactiveValues list instead of input$categorical_var when saving a filter from edit mode.