Attempt: control which plots spinner reacts to
Added plots_tab_spinner parameter to the plots module. When this is TRUE, the module sends a message to JavaScript upon initialization, telling the waiter spinner for the plots tab to show when the plot in the module is re-computed. The message contains the full ID of the plot, which is stored in a JavaScript array. The current attempt does not work: no UI elements render at all.