Fix Interface Issue with PickerInput menus
The pickerInput menu for selecting filters based on categorical metadata extends beyond the width of the sidebar panel when very long metadata values are present. This causes the menu to appear behind the main panel, and a scrollbar appears in the sidebar panel on the x-axis. When items are selected, the checkmarks also appear at the right edge of the panel, and are not visible unless the user scrolls the entire sidebar panel.

CSS is added to enforce a max-width attribute for the items in the pickerInput menu. For very long items, the pickerInput itself will scroll on the x-axis, instead of the entire sidebar panel. Checkmarks now appear at the edge of the max width: while they display over long options, they are no longer behind the main panel.