Main app loading code for anndata objects
Added code to load objects based on the extension of the file path provided. .rds and .h5ad are currently supported. For SCE objects with  DelayedArray matrices, the loading code is still determined via the is_HDF5SummmarizedExperiment value in the object-specific config file. Extensions are checked when this value is FALSE, and the loadHDF5SummarizedExperiment() function is ran when this value is TRUE.