Remove unnecessary reactive variable from plots_tab
With the current implementation of the filter menus, there is no need for the object_init() variable in the plots tab. the object_init() variable was previously necessary to get the subset reactive expression to return the full object when the object is loaded, but this can now happen without this variable as long as the subset reactive expression responds to a change in the object. The output of the subset_selections module was previously undefined when an object was loaded, which caused an error when passing the output to make_subset(). The output of this module is now the same as the output (an empty list) when the full object is requested, which will direct the subset reactive statement to properly return the full object.

Removing this now-unnecessary reactive variable simplifies the app and reduces the likelihood of unexpected bugs when adding new features to the app.