Restore scroll position of plots tab window after spinner is removed
Addresses #222. The plots tab window must be hidden while a spinner is shown to keep the user from being able to scroll beneath the spinner, but this originally resulted in the scroll position of the window being lost. This is now restored by use of javascript functions, via shinyjs, that record the scroll position of the window when the spinner is shown, and restore the position after the spinner is hidden.