Pre-compute Dataset Setup Article
The dataset setup article involves operations on an object that is too big to host with the package on GitHub. The Rmd for the article must be evaluated by Github Actions to create the Pkgdown site, so a solution is necessary to get the Rmd to evaluate without having to post the data to github.

This was achieved by pre-computing the vignette, using the process detailed at https://ropensci.org/blog/2019/12/08/precompute-vignettes/

knitr::knit was ran to generate a Rmd with included code chunks and associated outputs. Plots generated were moved to the vignettes directory and renamed, and the SASS code chunk was manually added back into the output Rmd.