Manually correct pre-computed vignette
The DT datatables in the dataset setup walkthrough were being rendered as images instead of HTML upon pre-computation, and the images were being rendered incorrectly. Since the code for the datatables does not depend on the object in the vignette, it was added back in to the pre-computed vignette, for evaluation at the time the pkgdown site is created.