Fix #313
The cause of #313 is the variable_length_custom_title() reactive expression in the plot module. The reactive expression either returns a title value to use on plots, or NULL to use defaults. Originally, the expression was NULL whenever the title settings menu was equal to anything besides "custom". The default value was equal to the name of the feature when split_by was undefined on feature plots. When #305 was introduced, a toggle to display a "raw" feature name without the modality key vs. one with the key was added, and code was added to variable_length_custom_title to define a title value to use for the feature name based on whether the user requested raw feature names or not. The issue was that the title on feature plots is not always equal to a feature name: when a split_by variable is defined, it is equal to the names of the split_by groups. This unexpectedly introduced #313. To fix the issue, variable_length_custom_title was modified to return a different title value based on whether the title depends on a feature name, or on something else. The feature_based_title variable was created and conditionally set to TRUE or FALSE based on the settings entered. When TRUE, variable_length_custom_title returns feature names with the modality keys removed or kept based on whether "raw features" are enabled, and when FALSE, NULL is returned, directing plot_feature to use default title values. In the case of a split feature plot, this is equal to the names of the split by groups.