Fix error in simple groups dge
The subset filters were not properly formed in the case where DGE was performed using groups. This was found to be due to issues with return() statements in the new conditional statements in forming the filters. The statements have been corrected to properly return the subset filters in all cases.