Fix issue with GitHub actions
A live code chunk was inserted with the changes in #351, which caused the GitHub runner to throw an error. `dataset_setup_walkthrough.Rmd` should not contain executing R code since the dataset in this vignette can't be processed by the GitHub runner. This vignette should be markdown only

eval=FALSE was added, which resolved the error.