Modify config app code in response to #359
The issue in #359 was triggered by the metadata_type function, which was written with the expectation that metadata variables would only have one class. The function was re-written to account for the possibility of multiple variables, using inherits(), which is better practice regardless.

`metadata_type` also only gives the display name for a metadata class in the app, and was being used downstream to execute server code based on the class of metadata. It is not ideal to use a display name for a class to do this, so code in the options module was also re-factored to test for class independent of this function.