Fix error with AppDriver for shinytest2
When running AppDriver, shinyjs functions do not initialize properly, which caused an error in an observe statement in the plot module which crashes the app. The reason for this is not known, but the error is now prevented by adding a conditional statement that runs only when the shinyjs function is defined.