Dynamically re-size app body
Previously, the heights of the sidebar and main panels of the app were set to 85% of the window height, which resulted in white space at the bottom of the screen that is visually unappealing and limits the space available for the user to interact with the UI.

Now, the heights of these elements are set to precisely the height of the window minus the size of the navbar and spacing, previously set by the dynamicResize function.

The body of the app now extends to the bottom of the screen, without white space at the bottom.