Remove metadata stats output in DGE tab
The stats showing unique values of a subset in the DGE tab caused performance issues when a large object was loaded with many unique values in some metadata variables. The interface in its current form is cluttered, especially when many metadata variable are present, and it provides questionable value as a DGE test summary. Because of this, the interface was removed entirely.