Add collapsible sidebar to config app
The current config interface is difficult to use due to the drag and drop interface for metadata and reductions taking up most of the screen, making it difficult to enter options. A collapsible sidebar from bslib was added to address this. Admins can now expand and contract the sidebar to more easily access the options interface for assays, metadata, and reductions.