Add plotting functions from SCUBA
The plot_functions from SCUBA and any dependent functions have been moved to scExploreR, as they are more appropriate as part of scExploreR than as part of SCUBA.

A generic was also added, relabel_axis, which called by expr_plot. A method was added to this generic for MuData objects.